pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chatterbot"
version = "0.1.0"
description = "Silent Twitch chatter-profiling sidecar bot. Streamer-only profile surface; no chat-facing read path."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "twitchio>=2.8.0,<3.0.0",
    "httpx>=0.25.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "sqlite-vec>=0.1.6",
    "textual>=0.60.0",
    "python-socketio[asyncio_client]>=5.10.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
chatterbot = "chatterbot.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/chatterbot"]

[tool.ruff]
line-length = 100
target-version = "py311"

[dependency-groups]
dev = [
    "playwright>=1.58.0",
]
